Refactored Backend Code. Added simple explaination to README.md within backend directory.
