+++ backend/main.py
@@ -1,7 +1,12 @@
+import sys
+
 from fastapi import FastAPI, Depends
 from fastapi.middleware.cors import CORSMiddleware
-# from .routers import elastic_router, mysql_router, mongo_router, keyword_router
+
 from .routers import elastic_router, mysql_router, mongo_router, review, upload, keywordExt_router
+
+
+sys.stdout.reconfigure(encoding='utf-8')
 
 app = FastAPI()
 
@@ -35,5 +40,3 @@ app.include_router(mongo_router.router)
 app.include_router(review.router)
 app.include_router(upload.router)
 app.include_router(keywordExt_router.router)
-# app.include_router(keyword_router.router, prefix="/api")
-
