+++ backend/requirements.txt
@@ -14,3 +14,4 @@ scipy
 python-multipart
 passlib[bcrypt]
 email-validator
+pillow
