+++ backend/schemas/photo.py
@@ -1,4 +1,3 @@
-
 from pydantic import BaseModel
 from typing import List
