+++ backend/scripts/elastic/delete.py
