+++ backend/scripts/elastic/import_restaurants.py
